+++ src/Main.h
@@ -6,4 +6,4 @@ extern HWND ghWnd;
 extern BOOL bFullscreen;
 
 void PutFramePerSecound(void);
-int GetFramePerSecound(void);
+unsigned long GetFramePerSecound(void);
